% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerGeneFinder.R
\name{CancerGeneFinder}
\alias{CancerGeneFinder}
\title{Cancer-associated gene identification}
\usage{
CancerGeneFinder()
}
\description{
CancerGeneFinder is a CMC-based model to identify the genes whose mutations are responsible for a particular cancer initiation and progression.
}
