\name{CMC-package}
%\alias{CMC-package}
\alias{CMC}
\docType{package}
\title{
Cronbach-Mesbach Curve
}
\description{
This package calculates and plots the step-by-step Cronbach-Mesbach curve, that is a method, based on the Cronbach alpha coefficient of reliability, for checking the unidimensionality of a measurement scale.
}
\details{
\tabular{ll}{
Package: \tab CMC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Michela Cameletti and Valeria Caviezel

Maintainer: Michela Cameletti \email{ michela.cameletti@unibg.it }
}

\references{
Curt, F., Mesbah, M., Lellouch, J. and Dellatolas, G. (1997) \emph{Handedness scale how many and which items?} Laterality, 2, 137--154.

Hamon, A. and Mesbah, M. (2002) \emph{Questionnaire reliability under the Rasch model}. Statistical Methods for Quality of Life Studies: Design, Measurement and Analysis. Mesbah, M., Cole, B.F. and Lee, M.L.T. (Eds.), Kluwer Academic Publishing, Boston, 155--168.

Mesbah, M. (2010) \emph{Statistical quality of life}. Method and Applications of Statistics in the Life and Health Sciences. Balakrishnan, N. (Editor), Wiley, 839--864.

Nordmann, J., Mesbah, M., Berdeaux, G. (2005) \emph{Scoring  of Visual Field Measured through Humphrey Perimetry: Principal Component Varimax Rotation Followed by Validated Cluster Analysis}. Investigative Ophthalmology & Visual Science 46, 3169--3176.
}

\keyword{ package }
