% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMC_R.R
\name{CMC}
\alias{CMC}
\title{Conditional Multifactorial Contingency (CMC) model}
\usage{
CMC(tensor,Y=1,items_used=NULL)
}
\arguments{
\item{tensor}{a high dimension data array}

\item{Y}{(Optional) Indicate the number of trials of each entries of tensor. The input should be an array with the same size as tensor, or a scalar. The latter case implies that all entries share the same value of Y. If tensor is binary, Y should be equal to 1. By default, Y = 1.}

\item{items_used}{(Optional) Use when there are missing values in tensor. The input should be a binary array with the same size as tensor, a scalar value of 1, or NULL. The latter two cases imply that there are no missing value in tensor. If there are missing values in tensor, then the corresponding entries in items_used should be 0, otherwise, 1. By default, items_used = NULL.}
}
\description{
The observed data are often influenced by numerous intrinsic or extrinsic factors at the same time, which may distort or even bury the impact of variables of real interest. Explicitly modeling the effects of a large number of factors is challenging, as no reliable prior  knowledge is available on how these factors exert their effects, as well as to what degree the effects are. What's more, the numerous factors' effects are often intertwined, and call for joint consideration of all the factors.  Conditional Multifactorial Contingency (CMC) is a model to jointly learn the multifactorial effect in large-scale data. Specifically, it reformulates the data as a contingency tensor with each dimension corresponding to a factor class, and then systematically models and learns the multiple factors as well as the joint probability distribution conditional on these factor.
}
\author{
Zuolin Cheng
}
