% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc-downstream.R
\name{construct.cmc}
\alias{construct.cmc}
\alias{construct}
\alias{decompose.cmc}
\alias{decompose}
\alias{predict.cmc}
\title{Add structure to a null model}
\usage{
\method{construct}{cmc}(object, newdata = NULL, from_col, to_col)

construct(object, newdata = NULL, from_col, to_col)

\method{decompose}{cmc}(object, newdata = NULL, from_col, to_col)

decompose(object, newdata = NULL, from_col, to_col)

\method{predict}{cmc}(object, ..., newdata = NULL, what = "pdist", at,
  to_col)
}
\arguments{
\item{object}{A cmc-fitted model.}

\item{newdata}{Data frame to operate on}

\item{from_col}{Name of column containing the null values. For example,
probabilities (PIT scores) to evaluate the quantile functions at.}

\item{to_col}{Name of the column to append the output to. Leave blank if
you want a vector output.}

\item{...}{Not used}

\item{what}{What to predict. Could be "pdist", "qdist", or "rdist".}

\item{at}{Vector of values to evaluate the \code{what} function at.}
}
\description{
Converts marginal Unif(0,1) distributions to have modelled distributions.
CURRENTLY ONLY HANDLES PIT SCORES AS VALUES.
}
\details{
If from_col values are numbers, these are interpreted as PIT
scores by applying the quantile function corresponding
to each predictive distribution.
If values are distributions, these are converted in
such a way that the predictive distributions are obtained if the values
are all Unif(0,1) distributions.
}
