% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_mle_raw}
\alias{cmc_mle_raw}
\title{Fit CMC model via CNQR or MLE}
\usage{
cmc_mle_raw(udat, force_ig, method, xvine, u2cond)
}
\arguments{
\item{udat}{matrix of PIT score data, in the order of response (V),
first predictor (U1), second predictor (U2), linked by V-U1-U2.}

\item{force_ig}{TRUE if you want (U1, V) ~ IG copula.}
}
\value{
Two vines: vine1 is the copula linking (Y, X1), and
vine2 is the copula linking (Y, X2)|X1.
}
\description{
Provides a raw workhorse for CNQR (cmc_cnqr_raw) or MLE (cmc_mle_raw).
Need to provide it with highly processed input. Intended to be used
as internal functions.
}
