% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_mle_raw}
\alias{cmc_mle_raw}
\alias{cmc_cnqr_raw}
\title{Fit CMC model via CNQR or MLE}
\usage{
cmc_mle_raw(udat, force_ig, method, xvine, u2cond)

cmc_cnqr_raw(udat, force_ig, xvine, u2cond, sc, verbose, copspace = NULL,
  families = c("indepcop", "bvncop", "bvtcop", "mtcj", "gum", "frk",
  "joe", "bb1", "bskewncop", "bskewncopp"))
}
\arguments{
\item{udat}{matrix of PIT score data, in the order of response (V),
first predictor (U1), second predictor (U2), linked by V-U1-U2.}

\item{force_ig}{TRUE if you want (U1, V) ~ IG copula.}

\item{method}{Passed to \code{fit_igcop_mle}.}

\item{xvine}{Copula joining X1 and X2, bundled into a vine object.}

\item{u2cond}{Vector of U2|U1.}

\item{sc}{Scorer, as in the output of \code{cnqr::scorer}}

\item{verbose}{Passed to \code{cnqr_reduced}}

\item{copspace}{Passed to \code{cnqr_reduced}. \code{force_ig}
takes precedence.}

\item{families}{Vector of copula family names acting as a "pool"
to choose from when fitting.}
}
\value{
Two vines: vine1 is the copula linking (Y, X1), and
vine2 is the copula linking (Y, X2)|X1.
}
\description{
Provides a raw workhorse for CNQR (cmc_cnqr_raw) or MLE (cmc_mle_raw).
Need to provide it with highly processed input. Intended to be used
as internal functions.
}
