% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnqr_reduced.R
\name{cnqr_reduced}
\alias{cnqr_reduced}
\title{Reduced version of cnqr::cnqr().}
\usage{
cnqr_reduced(edges, dat, sc, basevine, pdist = identity, QY = identity,
  copspace = NULL, refit = FALSE, verbose = FALSE,
  families = c("indepcop", "bvncop", "bvtcop", "mtcj", "gum", "frk",
  "joe", "bb1", "bskewncop", "bskewncopp"))
}
\description{
Only difference is that it doesn't try to compute a final score on the
fitted model.
}
