% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{compute_cmc_mle}
\alias{compute_cmc_mle}
\alias{compute_cmc_mle_igcop}
\alias{compute_cmc_cnqr}
\title{Fit CMC model via CNQR or MLE}
\usage{
compute_cmc_mle(v, u1, u2cond)

compute_cmc_mle_igcop(v, u1, u2cond, method)

compute_cmc_cnqr(udat, force_ig, xvine, u2cond, sc, verbose,
  copspace = NULL, families = c("indepcop", "bvncop", "bvtcop", "mtcj",
  "gum", "frk", "joe", "bb1", "bskewncop", "bskewncopp"))
}
\arguments{
\item{v}{Vector of PIT scores of the response}

\item{u1}{Vector of PIT scores of the first predictor to link to the response.}

\item{u2cond}{Vector of PIT scores of the second predictor, conditional on
the first.}

\item{method}{Passed to \code{fit_igcop_mle}.}

\item{force_ig}{TRUE if you want}

\item{xvine}{Copula joining the 1st and 2nd columns of dmat.}

\item{sc}{Scorer, as in the output of \code{scorer}}

\item{verbose}{Passed to \code{cnqr_reduced}}

\item{copspace}{Passed to \code{cnqr_reduced}. \code{force_ig}
takes precedence.}

\item{families}{Vector of copula family names acting as a "pool"
to choose from when fitting.}
}
\value{
Two vines: vine1 is the copula linking (X1, Y), and
vine2 is the copula linking (X2, Y)|X1.
}
\description{
Provides a raw workhorse for CNQR (compute_cmc_cnqr) or MLE (compute_cmc_mle).
Need to provide it with highly processed input. Intended to be used
as internal functions. The "_ig" version forces (U1, V) to have an IG copula.
}
