% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivingTFDetection_ChIPseq_Mouse.R
\name{DrivingTFDetection_ChIPseq_Mouse}
\alias{DrivingTFDetection_ChIPseq_Mouse}
\title{Driving TF detection}
\usage{
DrivingTFDetection_ChIPseq_Mouse(DEGs)
}
\arguments{
\item{DEGs}{a character vector of query gene set, such as a set of differential expression genes (DEGs)}
}
\description{
Detection of driving transcription factors (TFs) for a set of co-expression genes. The cell type specific TF binding sites are extracted from
   mouse ChIP-seq database. In this package, CMC model is utilized to systematically model the gene/TF heterogeneity in
   binding affinities, and enables a more powerful detection.
}
\examples{
DEGs_example_MYOD1 <- getTestGeneSet()
DrivingTFDetection_ChIPseq_Mouse(DEGs_example_MYOD1)
}
