% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc-downstream.R
\name{eval.cmc}
\alias{eval.cmc}
\title{Evaluate cmc model}
\usage{
\method{eval}{cmc}(object, newdata = NULL)
}
\arguments{
\item{object}{A cmc-fitted model.}

\item{newdata}{Data frame to operate on}
}
\description{
Computes things useful for downstream computations.
}
