% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePlot.R
\name{GenePlot}
\alias{GenePlot}
\title{GenePlot}
\usage{
GenePlot(List, Gene_list, Pt_size = 1)
}
\arguments{
\item{List}{The "Gotana" object. \cr}

\item{Gene_list}{Please check function "FeaturePlot" of "Seurat" package ("features").\cr}

\item{Pt_size}{Please check function "FeaturePlot" of "Seurat" package ("pt.size").\cr}
}
\value{
A Figure.\cr
}
\description{
This is the function that show the gene expression level in GO term analysis based on the UMAP of "Seurat_GO".\cr
}
\details{
This is some description of this function.

For "-log10_p" & "p_value" scores, this figure will show -log10(p-value).\cr
         For "Z_score", this figure will show adjusted Z-score. (Change the lower bound of the displayed data).\cr
}
