% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivingTFDetection_ChIPseq_Mouse.R
\name{getTestGeneSet}
\alias{getTestGeneSet}
\title{Gene set for test}
\usage{
getTestGeneSet()
}
\description{
Return a set of genes for test purpose. The gene set is the differential expression genes(DEGs) between
MYOD1-KO cells and controls. The KO experiment and DEGs analysis were conducted in embryonic fibroblasts.
}
\examples{
DEGs_example_MYOD1 <- getTestGeneSet()
}
