% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_Scores.R
\name{GO_Scores}
\alias{GO_Scores}
\title{GO_Scores}
\usage{
GO_Scores(List)
}
\arguments{
\item{List}{The "Gotana" object. \cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List" & “CMC” 
        & "GO_Term_Activity_Scores". \cr
        \cr
        In the sub-list "GO_Term_Activity_Scores", there are 3 variables:\cr
        \cr
        p_Value: The p-value (GO term activity score) of every cell in different GO terms.\cr
        \cr
        Z_Score: The Z-score (GO term activity score) of every cell in different GO terms.\cr
        \cr
        N_Gene: Numbers of genes considered in every GO term.\cr
        \cr
}
\description{
This is the function that calculate GO term activity scores (Z-score & p-value).
}
\details{
This is some description of this function.

For example:\cr
\code{
GO_analysis_example <- GO_Scores(GO_analysis_example)
}
}
