% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_selected_Order_Statistics.R
\name{GO_selected_Order_Statistics}
\alias{GO_selected_Order_Statistics}
\title{GO_selected_Order_Statistics}
\usage{
GO_selected_Order_Statistics(List)
}
\arguments{
\item{List}{The "Gotana" object. \cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List" & “CMC” 
        & "GO_Term_Activity_Scores" & "Feature_Selection". \cr
        \cr
        In the sub-list "Feature_Selection", there are 4 variables:\cr
        \cr
        Adjusted_p_Value: Adjusted_p_Value of GO terms; the smaller the value, the more significant the corresponding GO term.\cr
        \cr
        Adjusted_p_Value_sorted: Sorted adjusted_p_Value of GO terms.\cr
        \cr
        Index_Sort: Index of sorted GO terms in original List.\cr
        \cr
        GO_Term_sorted: Sorted GO term list.
        \cr
}
\description{
This is the function that select significant GO terms.
}
\details{
This is some description of this function.

This function use order statistics test of standard normal distribution.\cr
         After running the function, the histogram figure of adjusted_p_Value in all GO terms will be shown.\cr
}
