% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_GOSet.R
\name{Map_GOSet}
\alias{Map_GOSet}
\title{Map_GOSet}
\usage{
Map_GOSet(List)
}
\arguments{
\item{List}{The "Gotana" object. \cr}
}
\value{
An object including sublist "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List".\cr
        In the sub-list "GO_Dataset", there are 3 variables:\cr
        GO_Term_list: A table that records the GO terms' ID & Description. (Built-in dataset of package)\cr
        \cr
        Gene_list: A list that saves the whole gene list of the GO term dataset. (Built-in dataset of package)\cr
        \cr
        Map: A binary matrix; rows mean GO terms and columns mean genes,Showing the mapping relationship between GO terms and genes.\cr
        (The orders are consistent with the data matrix, Built-in dataset of package)\cr
        \cr
        In the sub-list "AfterMapping_List", there are 6 variables:\cr
        Size: Numbers of genes, cells, and GO terms in intersection of the sample dataset and the GO term dataset.\cr
        \cr
        Data: Intersection data of the sample dataset and the GO term dataset.\cr
        \cr
        Data_Bin: Binary transformation result of Data.\cr
        \cr
        Gene: Intersection gene list of the sample dataset and the GO term dataset.\cr
        \cr
        GO_Term_Filted: Intersecting GO term table including ID & Description.\cr
        \cr
        Map: Intersecting map matrix.\cr
}
\description{
This is the function that map the sample data to GO term dataset.
}
\details{
This is some description of this function.

Integrate the GO term dataset into the "GOtana" object. \cr
         Get the intersection of the sample dataset and the gene list in the GO term dataset.
}
