% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers_Detection.R
\name{Markers_Detection}
\alias{Markers_Detection}
\title{Markers_Detection}
\usage{
Markers_Detection(
  Object,
  cell_group1,
  cell_group2,
  Flag_testAllgene = FALSE,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  only.pos = FALSE
)
}
\arguments{
\item{Object}{The "Gotana" object. \cr}

\item{cell_group1}{Name list of cells in group1.\cr}

\item{cell_group2}{Name list of cells in group2.\cr}

\item{Flag_testAllgene}{Default: FALSE. \cr}

\item{min.pct}{Please check function "FindMarkers" of "Seurat" package.\cr}

\item{logfc.threshold}{Please check function "FindMarkers" of "Seurat" package.\cr}

\item{only.pos}{Please check function "FindMarkers" of "Seurat" package.\cr}
}
\value{
An DEGs list of cell_group1.\cr
}
\description{
This is the function that detect DEGs of "cell_group1", comparing "cell_group1" with "cell_group2".\cr
}
\details{
This is some description of this function.

The input of parameters "cell_group1" & "cell_group2" should be name list of cells in these two group.\cr
         For example: A "Gotana" example named "GO_analysis_example":\cr
         \cr
         Cluster_Results <- GO_analysis_example$AfterMapping_List$Cluster\cr
         Cluster_0 <- GO_analysis_example$AfterQC_List$Cell_ID[Cluster_Results == '0']\cr
         Cluster_1 <- GO_analysis_example$AfterQC_List$Cell_ID[Cluster_Results == '1']\cr
         Cluster_2 <- GO_analysis_example$AfterQC_List$Cell_ID[Cluster_Results == '2']\cr
         Markers_Cluster0 <- Markers_Detection(GO_analysis_example, Cluster_0, c(Cluster_1,Cluster_2), 
                                              Flag_testAllgene = FALSE, min.pct = 0.1, logfc.threshold = 0.25, only.pos = TRUE)\cr
}
