% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscoreplot.R
\name{nscoreplot}
\alias{nscoreplot}
\title{Quick Normal Scores Plot}
\usage{
nscoreplot(.data, x_col, y_col)
}
\arguments{
\item{.data}{Data frame containing the data.}

\item{x_col, y_col}{Names of the columns containing the
X and Y variables, respectively.}
}
\description{
Produces a quick base R normal scores plot.
}
