% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_scData.R
\name{QC_scData}
\alias{QC_scData}
\title{QC_scData}
\usage{
QC_scData(
  List,
  Gene_threshold = 0.01,
  Count_threshold = 0,
  Cell_threshold = 500,
  MT_threshold = 0
)
}
\arguments{
\item{List}{The "Gotana" object that was built after run "Read_scRNA" function. \cr}

\item{Gene_threshold}{The threshold for the percentage of all expressed genes, to remove some genes that have no expression in almost all cells. \cr}

\item{Count_threshold}{The threshold for total counts.\cr}

\item{Cell_threshold}{The threshold for the number of genes the cell expresses, to remove some cells that have low numbers of expressed genes. \cr}

\item{MT_threshold}{The threshold for the percentage of mitochondrial genes. \cr}
}
\value{
An object including sublist "Original_List" & "AfterQC_List".\cr
         \cr
        In the sub-list "AfterQC_List", there are 8 variables:\cr
        \cr
        Size:             Numbers of genes and cells in the data.\cr
        \cr
        Data:             A count matrix of the data.\cr
        \cr
        Gene:             Gene list of the data. (The order is consistent with the data matrix)\cr
        \cr
        Cell_ID:          Cell list of the data. (The order is consistent with the data matrix)\cr
        \cr
        Mean_Expression:  In a single cell, the percentage of all expressed genes in the gene list.\cr
        \cr
        Sum_Count:        In a single cell, the total counts of all expressed genes. (-log10)\cr
        \cr
        Sum_nFeature:     In a single cell, the number of genes the cell expresses.\cr
        \cr
        MT_Percent:       In a single cell, the count percentage of mitochondrial genes.\cr
}
\description{
This is the function that perform QC process.
}
\details{
This is some description of this function.

To set QC thresholds, all parameters will be set based on experience or those four histogram figures. \cr
         ("Expression Percentage", "Total Counts", "Expression Number", "MT Percent")
}
