% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_scData.R
\name{Read_scData}
\alias{Read_scData}
\title{Read_scData}
\usage{
Read_scData(Input_str, Str_mt = "mt-")
}
\arguments{
\item{Input_str}{Input_str is the input ".txt" file including its path (Like this: "./GO_term_analysis/scRNA_seq_Data.txt").\cr}

\item{Str_mt}{Str_mt is to use the set that all genes starting with string Str_mt as a set of mitochondrial genes. The default value is "mt-".\cr}
}
\value{
A List including sublist "Original_List".\cr
       \cr
        In the sub-list "Original_List", there are 8 variables:\cr
        \cr
        Size_orig:             Numbers of genes and cells in the data.\cr
        \cr
        Data_orig:             A count matrix of the data.\cr
        \cr
        Gene_orig:             Gene list of the data. (The order is consistent with the data matrix)\cr
        \cr
        Cell_ID_orig:          Cell list of the data. (The order is consistent with the data matrix)\cr
        \cr
        Mean_Expression_orig:  In a single cell, the percentage of all expressed genes in the gene list.\cr
        \cr
        Sum_Count_orig:        In a single cell, the total counts of all expressed genes. (-log10)\cr
        \cr
        Sum_nFeature_orig:     In a single cell, the number of genes the cell expresses.\cr
        \cr
        MT_Percent_orig:       In a single cell, the count percentage of mitochondrial genes.\cr
}
\description{
This is the function that read the original scRNA-seq data file.
}
\details{
This is some description of this function.

The formats of the input file is ".txt" or ".csv". Row names are gene names, and column names are cell IDs.\cr
         For example:\cr
  Gene  \\  Cell_1  \\  Cell_2  \\  Cell_3    ... \cr
  Gene_1  \\  0     \\     1    \\    2       ... \cr
  Gene_2  \\  1     \\    4     \\    6       ... \cr
  ...  \\    ...   \\    ...   \\    ...      ... \cr
  (First row in the ".txt" file should be like this: text "Gene" + cell IDs)\cr
  (First column should be like this : text "Gene" + gene names)\cr
}
