% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Target_Genes.R
\name{Read_Target_Genes}
\alias{Read_Target_Genes}
\title{Read_Target_Genes}
\usage{
Read_Target_Genes(List, Input_str)
}
\arguments{
\item{List}{The "TySim" object. \cr}

\item{Input_str}{Input_str is a ".txt" or ".csv" file (Target DEGs list) including its path (Like this: "./Similarity_analysis/"DEGs_example.txt").\cr}
}
\value{
An object including sublist "Original_List" & "AfterQC_List" & "Target_Genes".\cr
         \cr
}
\description{
This is the function that perform QC process.
}
\details{
This is some description of this function.

To set QC thresholds, all parameters will be set based on experience or those four histogram figures. \cr
         ("Expression Percentage", "Total Counts", "Expression Number", "MT Percent")
}
