% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_CMC.R
\name{Run_CMC}
\alias{Run_CMC}
\title{Run_CMC}
\usage{
Run_CMC(List)
}
\arguments{
\item{List}{The "TySim" object. \cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "Target_Genes" & “CMC”. \cr
         \cr
        In the sub-list "CMC", there are 7 variables:\cr
        \cr
        P_Out_Bin: the probability of every single entry in binary model.\cr
        \cr
        P_Out: the probability of every single entry in non-binary model.\cr
        \cr
        Expectation: the expectation of every single entry in non-binary model.\cr
        \cr
        Gene_Factor_Bin: the "sensitivity" of gene factor in binary model.\cr
        \cr
        Gene_factor: the "sensitivity" of gene factor in non-binary model.\cr
        \cr
        Cell_Factor_Bin: the "sensitivity" of cell factor in binary model.\cr
        \cr
        Cell_factor: the "sensitivity" of cell factor in non-binary model.\cr
        \cr
}
\description{
This is the function that run CMC Model.
}
\details{
This is some description of this function.

Build a null distribution of the sample data.
}
