% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_PCA.R
\name{Run_PCA}
\alias{Run_PCA}
\title{Run_PCA}
\usage{
Run_PCA(List, Score_type = "p_value", Selected_GOs = -1, Show_npcs = 30)
}
\arguments{
\item{List}{The "Gotana" object. \cr}

\item{Score_type}{Defaults:"p_value"; "p_value" or "Z_score" or "-log10_p".\cr}

\item{Selected_GOs}{The number of selected top significant GO terms (Integer). \cr
Defaults: -1; if (-1), select the GO terms that have: $Feature_Selection$Adjusted_p_Value_sorted < 0.05.\cr}

\item{Show_npcs}{The number of PCAs shown in Variance plot.}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List" & “CMC” 
        & "GO_Term_Activity_Scores" & "Feature_Selection" & "PCA". \cr
        \cr
        In the sub-list "PCA", there are 4 variables:\cr
        \cr
        PCA_matrix: PCA matrix.\cr
        \cr
        Variance: Variance of each PC.\cr
        \cr
        Cumulate_Percent: Sum of variance percentages in top PCs.\cr
        \cr
        Score_Type: Input parameter "Score_type".\cr
        \cr
        In the sub-list "Feature_Selection", "Selected_GOs" will be added in the sub-list.\cr
}
\description{
This is the function that run PCA.
}
\details{
This is some description of this function.

Users can select different GO term activity scores to perform PCA (p_value, Z_score, -log10_p).
}
