% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TySim.R
\name{TySim}
\alias{TySim}
\title{TySim}
\usage{
TySim(List)
}
\arguments{
\item{List}{The "TySim" object. \cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "Target_Genes" & “CMC” & "Similarity_Scores". \cr
        \cr
        In the sub-list "Similarity_Scores", there are 4 variables:\cr
        \cr
        p_Value: The p-value (similarity score) of every cell for the target DEGs (Combine binary and non-binary model).\cr
        \cr
        Z_Score: The Z-score (similarity score) of every cell for the target DEGs (Combine binary and non-binary model).\cr
        \cr
        Minus_log10_p_Value: -log10(p-value).\cr
        \cr
        Adjusted_Z_Score: Change the lower bound of the displayed Z-score.\cr
        \cr
}
\description{
This is the function that calculate similarity scores (Z-score & p-value).
}
\details{
This is some description of this function.

For example:\cr
\code{
Similarity_analysis_example <- Similarity_Scores(Similarity_analysis_example)
}
}
