% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Seurate.R
\name{Update_Seurate}
\alias{Update_Seurate}
\title{Update_Seurate}
\usage{
Update_Seurate(List, Seurat_Object)
}
\arguments{
\item{List}{The "Gotana" object. \cr}

\item{Seurat_Object}{The "Seurat" object.\cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List" & “CMC” 
        & "GO_Term_Activity_Scores" & "Feature_Selection" & "PCA" & "Seurat_GO" & "Seurat_Gene" & "UMAP". \cr
        \cr
        In the sub-list "AfterMapping_List", "Cluster" will be added in the sub-list.\cr
        In the sub-list "UMAP", UMAP reduction results in "Seurat" function will be added in the sub-list.\cr
}
\description{
This is the function that update "Gotana" object.\cr
}
\details{
This is some description of this function.

This function is necessary to upload the clustering results from "Seurat" analysis.\cr
}
