% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc}
\alias{cmc}
\title{Fit a CMC model}
\usage{
cmc(ycol, x1col, x2col, data, method = "optim", force_ig = TRUE,
  marginal = NULL, sc, verbose = FALSE, copspace = NULL,
  omit_na = TRUE, families = c("indepcop", "bvncop", "bvtcop", "mtcj",
  "gum", "frk", "joe", "bb1", "bskewncop", "bskewncopp"))
}
\arguments{
\item{ycol, x1col, x2col}{Character names of the columns.}

\item{data}{Data frame of data}

\item{method}{Method of fitting. For MLE (the default), one of
"optim" or "nlm", depending
on the numerical optimizer you want to use. If "cnqr",
fits by CNQR, and you need
to specify the sc argument.}

\item{force_ig}{Force Y and X1 to have dependence described by an IG copula?
Default is TRUE.}

\item{marginal}{Marginal model of class "dst" from the distplyr package.
Distribution is assumed to be the same for the response and
predictors. If NULL, assumes variables are already PIT scores.}

\item{sc}{Scorer obtained through scorer()}

\item{verbose}{Only works for CNQR. If TRUE, will output the fitting
status of CNQR.}

\item{copspace}{Only works for CNQR.}

\item{omit_na}{Omit rows of data where at least one of ycol, x1col,
and x2col are NA? Currently only TRUE works.}

\item{families}{Vector of copula family names acting as a "pool"
to choose from when fitting.}
}
\description{
Intended to be used after a marginal model has been fit with
composite_dist(). So, the actual fitting that's done here is
strictly copula fitting, but the output is the entire model.
Limited capability here -- two predictors are assumed to be
lags, and therefore having the same marginals.
}
