% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnqr_reduced.R
\name{cnqr_reduced}
\alias{cnqr_reduced}
\title{Reduced version of cnqr::cnqr().}
\usage{
cnqr_reduced(edges, dat, sc, basevine, pdist = identity, QY = identity,
  copspace = NULL, refit = FALSE, verbose = FALSE,
  families = c("indepcop", "bvncop", "bvtcop", "mtcj", "gum", "frk",
  "joe", "bb1", "bskewncop", "bskewncopp"))
}
\arguments{
\item{edges}{Integer vector starting with the column number of the response
variable, followed by the column numbers of the predictors to use,
in the order that they are to be linked with the response.}

\item{dat}{Data frame or matrix containing the data (columns=variables), or
a list of such data frames. In the latter case, the first data frame is
used for parameter estimation (training data), the second data
frame is used for copula selection (validation data), and all other
data are not used in the fitting procedure, but are included in the
output. \code{dat} may
also be a single vector of response data, in the case you
have no predictors.}

\item{sc}{Scoring rule to use for the regression, as in the output
of \code{cnqr::scorer}.}

\item{basevine}{Object of type \code{'rvine'} containing
the predictors, and not the response. If left blank, the predictors are
assumed to be independent.}

\item{pdist}{List of vectorized distribution functions of the data, where
the entries correspond respectively to the columns in the data frames in
\code{dat}. Or, if the distribution functions are all the same, \code{cdf}
can be that single function. You can ignore this argument if you don't
include any predictors in \code{edges}.}

\item{QY}{Quantile function of the response, which accepts a
vector of values (quantile levels) in (0,1). It should return
quantiles, either in the form of
a vector corresponding to the input,
or in the form of a matrix with columns corresponding to the inputted
quantile levels and rows corresponding to the observations
(thus allowing for each observation of the response to come from different
distributions).}

\item{copspace}{List with vector entries of the copula families to try fitting
for each edge. \code{NULL} entries will be replaced with all copula families
in \code{families}. Or, leave the argument as \code{NULL} to fit all families in
\code{families} for all edges.}

\item{refit}{After all the copula models have been selected and fit
(this is done sequentially), should the parameters of those copula
families be re-estimated, but this time altogether, using all the data?
\code{TRUE} if so.}

\item{verbose}{Logical; should messages be output to indicate what
\code{cnqr} is doing?}

\item{families}{Vector of copula families.
For those edges in \code{cop} that don't have copula families
specially selected (i.e. have \code{NULL} entries), these families are used.}
}
\description{
Only difference is that it doesn't try to compute a final score on the
fitted model.
}
