% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Seurat_Object.R
\name{Create_Seurat_Object}
\alias{Create_Seurat_Object}
\title{Create_Seurat_Object}
\usage{
Create_Seurat_Object(List)
}
\arguments{
\item{List}{The "Gotana" object. \cr}
}
\value{
An object including sub-list "Original_List" & "AfterQC_List" & "GO_Dataset" & "AfterMapping_List" & “CMC” 
        & "GO_Term_Activity_Scores" & "Feature_Selection" & "PCA" & "Seurat_GO" & "Seurat_Gene". \cr
        \cr
        In the sub-list "AfterQC_List", "Data_Normalized" will be added in the sub-list.\cr
}
\description{
This is the function that build Seurat object.\cr
}
\details{
This is some description of this function.

There are two Seurat objects.\cr
         In "Seurat_GO", the count matrix will be replaced by GO term activity score matrix. (Score type depends on the "Score_type" in Run_PCA)\cr
         In "Seurat_Gene", the object just perform "Normalization" process. (normalization.method = "LogNormalize", scale.factor = 100000)\cr
}
