% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc-downstream.R
\name{eval.cmc}
\alias{eval.cmc}
\title{'augment' only predicts on the training data.}
\usage{
\method{eval}{cmc}(object, newdata = NULL)
}
\arguments{
\item{tau}{Vector of quantile levels to make predictions at}
}
\description{
'augment' only predicts on the training data.
}
