% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_igcop.R
\name{fit_igcop_mle}
\alias{fit_igcop_mle}
\title{Fit IG copula via MLE or CNQR}
\usage{
fit_igcop_mle(u, v, method = "nlm", init = c(2, 2), ...)
}
\arguments{
\item{u, v}{Vectors of PIT scores}

\item{method}{"nlm" and "optim" for MLE (representing the name of the
numerical optimizer); or "cnqr" for CNQR.}

\item{init}{Vector of initial copula parameter values for use in
the numerical optimizer.}

\item{...}{Other arguments to pass to numerical optimizer}
}
\description{
Fit IG copula via MLE or CNQR
}
